% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{AggregatePeakCounts}
\alias{AggregatePeakCounts}
\title{Aggregate multiple peak count outputs together}
\usage{
AggregatePeakCounts(peak.sites.file, count.dirs, output.dir, exp.labels = NULL)
}
\arguments{
\item{peak.sites.file}{a file containing peak site coordinates}

\item{count.dirs}{a list of output directories from CountPeaks}

\item{output.dir}{output directory for aggregate count matrix}

\item{exp.labels}{optional labels to append to cell barcodes corresponding to count.dirs}
}
\value{
NULL. Writes counts to file.
}
\description{
Aggregate multiple peak count outputs together
}
\examples{

extdata_path <- system.file("extdata",package = "Sierra")
peak.sites.file <- paste0(extdata_path,"/TIP_merged_peaks.txt")
count.dirs <- c("example_TIP_sham_counts", "example_TIP_MI_counts")
output.dir <- "example_TIP_aggregate"

 \dontrun{ 
AggregatePeakCounts(peak.sites.file, count.dirs, output.dir)
}

}
