% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.R
\name{AnnotatePeaksFromGTF}
\alias{AnnotatePeaksFromGTF}
\title{Annotates a set of peak coordinates from a GTF}
\usage{
AnnotatePeaksFromGTF(peak.sites.file, gtf.file, output.file,
  genome = NULL, invert_strand = FALSE, annotationType = "any",
  transcriptDetails = TRUE, annotation_correction = TRUE,
  pA_motif_max_position = 50, AAA_motif_min_position = 10,
  polystretch_length = 13, max_mismatch = 1, append.chr.peaks = TRUE)
}
\arguments{
\item{peak.sites.file}{a file of peak coordinates.}

\item{gtf.file}{GTF reference file.}

\item{output.file}{file to write the annotations to.}

\item{genome}{genome object. If NOT NULL then will perform pA motif analysis.}

\item{invert_strand}{Boolean to signifiy if strand of gr peaks should be inversed}

\item{annotationType}{can be assigned "any" or "within". Default is "any" which states that the peak with gr must overlap annotation feature (eg exon)}

\item{transcriptDetails}{Boolean. If false will only return gene name. If true will return internal transcript position feature (eg exon/intron)}

\item{annotation_correction}{Boolean. When multiple overlapping genes are identified will prioritise gene based on annotation. 3'UTR annotation trumps all other annotation.}

\item{pA_motif_max_position}{Any AAUAAA after this position are not considered (default 50nt)}

\item{AAA_motif_min_position}{Any polyA/polyT stretches before this postion are not considered (default 10)}

\item{polystretch_length}{: the length of A or T to search for (default 13)}

\item{max_mismatch}{number of allowed mismatches for motif matching (default 1)}

\item{append.chr.peaks}{: When TRUE (default) appends the character "chr" on chromosome entry in peaks file.}
}
\value{
NULL. writes output to file
}
\description{
Annotate a set of peak coordinates according to genomic features the coordinates fall on -
3'UTR, exon, intron and 5'UTR, and annotate proximity to motifs. Motifs include the
canonical polyA motif, A-rich regions and T-rich regions.
}
\examples{
AnnotatePeaksFromGTF(peak.sites.file, gtf.file, output.file, genome)

}
