% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_usage.R
\name{apply_DEXSeq_test_sce}
\alias{apply_DEXSeq_test_sce}
\title{Apply DEXSeq to detect differential peak usage to a Single-Cell Experiment object}
\usage{
apply_DEXSeq_test_sce(peaks.sce.object, population.1,
  population.2 = NULL, exp.thresh = 0.1, fc.thresh = 0.25,
  adj.pval.thresh = 0.05, num.splits = 6, seed.use = 1,
  feature.type = c("UTR3", "UTR5", "exon", "intron"),
  filter.pA.stretch = FALSE, verbose = TRUE, do.MAPlot = FALSE,
  return.dexseq.res = FALSE, ncores = 1)
}
\arguments{
\item{peaks.sce.object}{SCE object of peaks}

\item{population.1}{a target population of cells (can be an ID/cluster label or a set of cell barcode IDs)}

\item{population.2}{comparison population of cells. If NULL (default), uses all non-population.1 cells}

\item{exp.thresh}{minimum percent expression threshold (for a population of cells) to include a peak}

\item{fc.thresh}{threshold for log2 fold-change difference for returned results}

\item{adj.pval.thresh}{threshold for adjusted P-value for returned results}

\item{num.splits}{the number of pseudo-bulk profiles to create per identity class (default: 6)}

\item{feature.type}{genomic feature types to run analysis on (degault: all)}

\item{filter.pA.stretch}{whether to filter out peaks annotated as proximal to an A-rich region (default: FALSE)}

\item{verbose}{whether to print outputs (TRUE by default)}

\item{return.dexseq.res}{return the raw and unfiltered DEXSeq results object (FALSE by default)}

\item{doMAPlot}{make an MA plot of results (FALSE by default)}
}
\value{
a data-frame of results.
}
\description{
Apply DEXSeq to detect differential peak usage been select populations. Works by building
a 'pseudo-bulk' profile of cell populations by aggregating counts from individual cells
into a smaller number of profiles, defined by num.splits.
}
\examples{
apply_DEXSeq_test(apa.seurat.object, population.1 = "1", population.2 = "2")


}
