% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{count_polyA}
\alias{count_polyA}
\title{count polyA sites in cells}
\usage{
count_polyA(polyA.sites.file, gtf.file, bamfile, whitelist.file,
  output.file, countUMI = TRUE, ncores = 1)
}
\arguments{
\item{polyA.sites.file}{a file containing polyA sites}

\item{gtf.file}{reference (GTF) file}

\item{bamfile}{scRNA-seq BAM file}

\item{whitelist.file}{white list file}

\item{output.file}{name of output}

\item{countUMI}{whether to count UMIs (default: TRUE)}
}
\value{
NULL. Writes counts to file.
}
\description{
Generates a count matrix
}
\examples{
count_polyA(polyA.sites.file, reference.file, bamfile, whitelist.file, output.file)

}
