% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{de_polya_lr}
\alias{de_polya_lr}
\title{Use the logistic regression method by Lior Patcher group to calculate differential polyA}
\usage{
de_polya_lr(apa.seurat.object, cluster1, cluster2, exp.thresh = 0.05,
  fc.thresh = 0.25, adj.pval.thresh = 1e-05, feature.type = c("UTR3",
  "UTR5", "exon", "intron"))
}
\arguments{
\item{apa.seurat.object}{a polyA Seurat object}

\item{cluster1}{foreground cluster}

\item{cluster2}{comparison cluster}

\item{exp.thresh}{threshold \% of cells expressing a peak to consider in}

\item{fc.thresh}{threshold log2 fold-change difference for testing a peak}

\item{adj.pval.thresh}{adjusted P-value threshold for retaining a peak}

\item{feature.type}{the feature type(s) to test for}
}
\value{
a data-frame of results
}
\description{
Use the logistic regression method by Lior Patcher group to calculate differential polyA
}
\examples{
res.table = lr_de_test(apa.seurat, cluster1 = "1", cluster2 = "2")

}
