% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{de_polya_MAST}
\alias{de_polya_MAST}
\title{Apply MAST differential expression test to polyA sites.}
\usage{
de_polya_MAST(apa.seurat.object, cluster1, cluster2, exp.thresh = 0.05,
  fc.thresh = 0.25, adj.pval.thresh = 1e-05)
}
\arguments{
\item{apa.seurat.object}{a polyA Seurat object}

\item{cluster1}{foreground cluster}

\item{cluster2}{comparison cluster}

\item{exp.thresh}{threshold \% of cells expressing a peak to consider in}

\item{fc.thresh}{threshold log2 fold-change difference for testing a peak}

\item{adj.pval.thresh}{adjusted P-value threshold for retaining a peak}
}
\value{
a list of P-values
}
\description{
Apply a MAST differential expression test to polyA sites, i.e. treats polyA sites as independent.
}
\examples{
find_MAST_de_polya(apa.seurat, cluster1 = "1", cluster2 = "2")

}
