% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_usage.R
\name{DetectUTRLengthShift}
\alias{DetectUTRLengthShift}
\title{Detect shifts in 3'UTR length usage between cell populations}
\usage{
DetectUTRLengthShift(
  peaks.object,
  gtf_gr,
  gtf_TxDb,
  population.1,
  population.2 = NULL,
  exp.thresh = 0.1,
  fc.thresh = 0.25,
  adj.pval.thresh = 0.05,
  num.splits = 6,
  seed.use = 1,
  verbose = TRUE,
  do.MAPlot = FALSE,
  ncores = 1
)
}
\arguments{
\item{peaks.object}{Either a Seurat or SCE object of peaks}

\item{gtf_gr}{GenomicRanges object from a GTF file}

\item{gtf_TxDb}{TxDb from gtf file}

\item{population.1}{a target population of cells (can be an ID/cluster label or a set of cell barcode IDs)}

\item{population.2}{comparison population of cells. If NULL (default), uses all non-population.1 cells}

\item{exp.thresh}{minimum percent expression threshold (for a population of cells) to include a peak}

\item{fc.thresh}{threshold for log2 fold-change difference for returned results}

\item{adj.pval.thresh}{threshold for adjusted P-value for returned results}

\item{num.splits}{the number of pseudo-bulk profiles to create per identity class (default: 6)}

\item{seed.use}{seed to set the randomised assignment of cells to pseudo-bulk profiles}

\item{verbose}{whether to print outputs (TRUE by default)}

\item{do.MAPlot}{make an MA plot of results (FALSE by default)}

\item{ncores}{Number of cores for multithreading}
}
\value{
a data-frame of results.
}
\description{
Detect global shifts in 3'UTR length usage between defined cell populations.
Firsts applies the DUTest function to detect differential usage (DU) peaks on 3'UTRs, 
after filtering out peaks annotated as proximal to A-rich regions. Identifies peaks
on the same 3'UTR as each DU peak, and determines a position of the DU peak on the
3'UTR relative to the terminating exon. Returns a table of DU results, with the location 
of each peak relative to the total number of peaks on the corresponding 3'UTR. Results 
table can be input to the PlotUTRLengthShift function to visualise the results, 
and evaluate global shifts.
}
