% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_merging.R
\name{do_peak_merging}
\alias{do_peak_merging}
\title{Merge peaks across a list of data-sets}
\usage{
do_peak_merging(apa.dataset.list, sim.thresh = 0.75,
  allow.match.var = 0.25)
}
\arguments{
\item{apa.dataset.list}{a named list containing peaks called from different data-sets}
}
\value{
a set of merged peaks
}
\description{
Takes as input a list of named peaks obtained from running peak calling on multiple data-sets.
First goes through each peak set and check what peaks within each set should be merged (self-merging).
Merging is based on similarity criteria set by sim.thresh and allow.match.var.
Then compares each peak set as a reference to the remaining sets to identify peaks that should be merged.
Returns a list of peaks that have been merged, as well as the unique peaks from each data-set.
}
\examples{
DoPeakMerging(apa.dataset.list)
}
