% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_usage.R
\name{DUTest}
\alias{DUTest}
\title{Apply DEXSeq to detect differential peak usage}
\usage{
DUTest(
  peaks.object,
  population.1,
  population.2 = NULL,
  exp.thresh = 0.1,
  fc.thresh = 0.25,
  adj.pval.thresh = 0.05,
  num.splits = 6,
  seed.use = 1,
  feature.type = c("UTR3", "exon"),
  filter.pA.stretch = FALSE,
  verbose = TRUE,
  do.MAPlot = FALSE,
  return.dexseq.res = FALSE,
  ncores = 1
)
}
\arguments{
\item{peaks.object}{Either a Seurat or SCE object of peaks}

\item{population.1}{a target population of cells (can be an ID/cluster label or a set of cell barcode IDs)}

\item{population.2}{comparison population of cells. If NULL (default), uses all non-population.1 cells}

\item{exp.thresh}{minimum percent expression threshold (for a population of cells) to include a peak}

\item{fc.thresh}{threshold for log2 fold-change difference for returned results}

\item{adj.pval.thresh}{threshold for adjusted P-value for returned results}

\item{num.splits}{the number of pseudo-bulk profiles to create per identity class (default: 6)}

\item{seed.use}{seed to set the randomised assignment of cells to pseudo-bulk profiles}

\item{feature.type}{genomic feature types to run analysis on (default: UTR3, exon)}

\item{filter.pA.stretch}{whether to filter out peaks annotated as proximal to an A-rich region (default: FALSE)}

\item{verbose}{whether to print outputs (TRUE by default)}

\item{do.MAPlot}{make an MA plot of results (FALSE by default)}

\item{return.dexseq.res}{return the raw and unfiltered DEXSeq results object (FALSE by default)}

\item{ncores}{number of cores to run DEXSeq with}
}
\value{
a data-frame of results.
}
\description{
Apply DEXSeq to detect differential peak usage been select populations. Works by building
a 'pseudo-bulk' profile of cell populations by aggregating counts from individual cells
into a smaller number of profiles, defined by num.splits.
}
\examples{



extdata_path <- system.file("extdata",package = "Sierra")
load(paste0(extdata_path,"/TIP_cell_info.RData"))
\dontrun{
peak.annotations <- read.table("TIP_merged_peak_annotations.txt", header = TRUE,sep = "\t",
                                      row.names = 1,stringsAsFactors = FALSE)
peaks.seurat <- NewPeakSeurat(peak.data = peak.counts, 
                             annot.info = peak.annotations, 
                             cell.idents = tip.populations,
                             tsne.coords = tip.tsne.coordinates,
                             min.cells = 0, min.peaks = 0)

res.table = DUTest(peaks.seurat, population.1 = "F-SL", population.2 = "EC1",
                         exp.thresh = 0.1,  feature.type = c("UTR3", "exon"))
}

}
