% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_de_genes}
\alias{find_de_genes}
\title{Apply differential expression testing to gene-level counts}
\usage{
find_de_genes(genes.seurat.object, cluster1, cluster2, exp.thresh = 0.05,
  fc.thresh = 0.25, adj.pval.thresh = 1e-05, test.use = "MAST")
}
\arguments{
\item{cluster1}{foreground cluster}

\item{cluster2}{comparison cluster}

\item{exp.thresh}{threshold \% of cells expressing a gene to consider in}

\item{fc.thresh}{threshold log2 fold-change difference for testing a gene}

\item{adj.pval.thresh}{adjusted P-value threshold for retaining a gene}

\item{test.use}{statistical test to use. Either MAST (default) or t-test.}

\item{apa.seurat.object}{a gene-level Seurat object}
}
\value{
a data-frame of results.
}
\description{
Essentially a wrapper function for find_de_polya. Applies the same approach for DE testing,
but without any reference to annotation information. Thus, this function can be applied to
a gene-level Seurat object to obtain differentially expressed genes for a direct comparison
with results from differential testing at the polyA level.
}
\examples{
find_de_genes(apa.seurat.object, cluster1 = "1", cluster2 = "2")

}
