% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_de_genes_collapsed_v2}
\alias{find_de_genes_collapsed_v2}
\title{Apply a differential expression test to gene-level collapsed polyA sites.}
\usage{
find_de_genes_collapsed_v2(apa.seurat.object, cluster1, cluster2,
  exp.thresh = 0.05, fc.thresh = 0.25, adj.pval.thresh = 1e-05,
  test.use = "MAST", feature.type = c("UTR3", "UTR5", "exon",
  "intron"), use.all.peaks = FALSE)
}
\arguments{
\item{apa.seurat.object}{a polyA Seurat object}

\item{cluster1}{foreground cluster}

\item{cluster2}{comparison cluster}

\item{exp.thresh}{threshold \% of cells expressing a peak to consider in}

\item{fc.thresh}{threshold log2 fold-change difference for testing a peak}

\item{adj.pval.thresh}{adjusted P-value threshold for retaining a peak}

\item{test.use}{statistical test to use. Either MAST (default) or t-test.}

\item{feature.type}{the feature type(s) to test for. Redundant if use.all.peaks set to TRUE.}

\item{use.all.peaks}{whether to use all peaks regardless of annotation. FALSE by default.}
}
\value{
a data-frame of results.
}
\description{
Collapses polyA sites into a single gene-level count for each gene. Performs differential
gene expression analysis on the collapsed gene-level counts.
}
\examples{
find_de_genes_collapsed(apa.seurat.object, cluster1 = "1", cluster2 = "2")

}
