% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_max_polyA_de}
\alias{find_max_polyA_de}
\title{Find genes showing differential usage of peak sites}
\usage{
find_max_polyA_de(apa.seurat.object, cluster1, cluster2 = NULL,
  exp.thresh = 0.1, fc.thresh = 0.5, peak.fc.thresh = 0.15,
  adj.pval.thresh = 1e-05, test.use = "t-test",
  feature.type = c("UTR3", "UTR5", "exon", "intron"),
  use.all.peaks = FALSE, add.annot.info = TRUE, print.output = TRUE)
}
\arguments{
\item{apa.seurat.object}{a polyA Seurat object}

\item{cluster1}{foreground cluster}

\item{cluster2}{comparison cluster}

\item{exp.thresh}{threshold \% of cells expressing a peak to consider in}

\item{fc.thresh}{threshold log2 fold-change difference for testing a peak}

\item{adj.pval.thresh}{adjusted P-value threshold for retaining a peak}

\item{test.use}{statistical test to use. Either MAST (default) or t-test.}

\item{feature.type}{the feature type(s) to test for. Redundant if use.all.peaks set to TRUE.}

\item{use.all.peaks}{whether to use all peaks regardless of annotation. FALSE by default.}

\item{add.annot.info}{whether to add genomic annotation to output (TRUE by default)}

\item{print.output}{whether to print the output}
}
\value{
a data-frame of results.
}
\description{
First run a differential expression test to identify peaks differentially expressed
between clusters. Selects genes that have at least one differential peak, but expressing
multiple peaks.
}
\examples{
find_de_polya(apa.seurat.object, cluster1 = "1", cluster2 = "2")

}
