% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{find_polyA}
\alias{find_polyA}
\title{Do peak calling on a scRNA-seq BAM file}
\usage{
find_polyA(output.file, reference.file, bamfile, junctions.file,
  min.jcutoff = 50, min.jcutoff.prop = 0.05, min.cov.cutoff = 500,
  min.cov.prop = 0.05, min.peak.cutoff = 200, min.peak.prop = 0.05,
  ncores = 1)
}
\arguments{
\item{output.file}{a file containing polyA sites}

\item{reference.file}{reference (GTF) file}

\item{bamfile}{scRNA-seq BAM file}

\item{junctions.file}{white list file}

\item{min.jcutoff}{name of output}

\item{min.jcutoff.prop}{whether to count UMIs (default: TRUE)}

\item{min.cov.cutoff}{min.cov.cutoff}

\item{min.cov.prop}{min.cov.prop}

\item{min.peak.cutoff}{min.peak.cutoff}

\item{min.peak.prop}{min.peak.prop}

\item{ncores}{number of cores to use}
}
\value{
NULL. Writes counts to file.
}
\description{
Do peak calling on a scRNA-seq BAM file...
}
\examples{
find_polyA(output.file, reference.file, bamfile, junctions.file)
}
