% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{FindPeaks}
\alias{FindPeaks}
\title{Do peak calling on a scRNA-seq BAM file}
\usage{
FindPeaks(output.file, gtf.file, bamfile, junctions.file,
  min.jcutoff = 50, min.jcutoff.prop = 0.05, min.cov.cutoff = 500,
  min.cov.prop = 0.05, min.peak.cutoff = 200, min.peak.prop = 0.05,
  ncores = 1)
}
\arguments{
\item{output.file}{a file containing polyA sites}

\item{gtf.file}{reference (GTF) file}

\item{bamfile}{scRNA-seq BAM file}

\item{junctions.file}{file of splice junctions (e.g. produced by regtools)}

\item{min.jcutoff}{minimum number of spliced reads across a junction for it to be considered (default: 50).}

\item{min.jcutoff.prop}{minimum proportion of junction reads out of all junction reads for that gene (default: 0.05)}

\item{min.cov.cutoff}{min.cov.cutoff}

\item{min.cov.prop}{min.cov.prop}

\item{min.peak.cutoff}{min.peak.cutoff}

\item{min.peak.prop}{min.peak.prop}

\item{ncores}{number of cores to use}
}
\value{
NULL. Writes counts to file.
}
\description{
Do peak calling on a scRNA-seq BAM file...
}
\examples{
FindPeaks(output.file, reference.file, bamfile, junctions.file)

}
