% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{fit_gaussian}
\alias{fit_gaussian}
\title{Fit Gaussian curve to the coverage}
\usage{
fit_gaussian(fit.data, maxval, fit.method, mu = 300)
}
\arguments{
\item{fit.data}{read coverage to fit}

\item{maxval}{maximum read coverage}

\item{fit.method}{Either NLS or MLE}

\item{mu}{initialised value for the centre of the peak (default: 300)}
}
\description{
Given read coverage data, fit a Guassian curve using either 
NLS or MLE fits.
}
