% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{gene_Labels}
\alias{gene_Labels}
\title{This function has been designed to be called from annotate_gr_from_gtf}
\usage{
gene_Labels(gr, reference_gr, annotationType)
}
\arguments{
\item{gr}{a granges object of peaks to annotate}

\item{reference_gr}{a granges object of annotation info}

\item{annotationType}{a granges object of peaks to annotate}
}
\description{
This function has been designed to be called from annotate_gr_from_gtf
}
