% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_merging.R
\name{generate_merged_peak_table}
\alias{generate_merged_peak_table}
\title{Merge peaks across data-sets based on a reference}
\usage{
generate_merged_peak_table(
  dataset.1,
  peak.dataset.list,
  self.merged.peaks.list,
  sim.thresh = 0.75,
  allow.match.var = 0.25,
  ncores = 1
)
}
\arguments{
\item{dataset.1}{the reference peak data-set}

\item{peak.dataset.list}{a list of peak data-sets}

\item{self.merged.peaks.list}{the set of self-merged peaks from the reference data-set}

\item{sim.thresh}{The required similarity threshold for merging (default: 0.75)}

\item{allow.match.var}{The allowance for deviation from the sim.thresh for comparison peaks (default: 0.25)}

\item{ncores}{Number of cores for multithreading}
}
\value{
a data-frame containing peaks, their class (merged or unique) and the original peak from the reference
}
\description{
Given a reference data-set, a list of data-sets for merging and set of merged peaks from the referece,
identify peaks that should be merged. Merged peaks are taken as the union of the peaks to be merged.
For two given peaks, A and B, they will be merged if at least one has some x\% (75\% by default) or more
overlap with the other, and the other has at least x-(y*x)\% overlap where y is a percentage of allowed
variance (25\% by default).
}
\examples{
\dontrun{
     generate_merged_peak_table(dataset.1, peak.dataset.table, self.merged.peaks.list)
 }
}
