% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_merging.R
\name{generate_self_merged_peaks}
\alias{generate_self_merged_peaks}
\title{Merge a set of peaks}
\usage{
generate_self_merged_peaks(apa.similarity.table, sim.thresh = 0.75,
  allow.match.var = 0.25)
}
\arguments{
\item{apa.similarity.table}{the set of peaks to merge}

\item{sim.thresh}{The required similarity threshold for merging (default: 0.75)}

\item{allow.match.var}{The allowance for deviation from the sim.thresh for comparison peaks (default: 0.25)}
}
\value{
a vector of merged peaks
}
\description{
Given a self-similarity table of peaks, identify peaks that should be merged. Merged peaks are
taken as the union of the two peaks. For two given peaks, A and B, they will be merged if at least one has
some x\% (75\% by default) or more overlap with the other, and the other has at least x-(y*x)\% overlap where
y is a percentage of allowed variance (25\% by default)
}
\examples{
generate_similarity_table(peaks.1)
}
