% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bams.R
\name{geneToGR}
\alias{geneToGR}
\title{geneToGR converts a gene symbol to genomic ranges coordinate}
\usage{
geneToGR(geneSymbol, gtf_gr)
}
\arguments{
\item{geneSymbol}{: Gene symbol}

\item{gtf_gr}{: Granges object of a gtf file}
}
\description{
geneToGR converts a gene symbol to genomic ranges coordinate
}
\examples{
    library('Sierra')
    extdata_path <- system.file("extdata",package = "Sierra")
    gtf.file <- paste0(extdata_path,"/Vignette_cellranger_genes_subset.gtf")
    gtf.gr <- rtracklayer::import(gtf.file)
    
    geneGR  <- geneToGR(geneSymbol= "Dnajc19",gtf_gr=gtf.gr)
}
