% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_usage.R
\name{get_expressed_peaks_sce}
\alias{get_expressed_peaks_sce}
\title{Identify highly expressed peaks}
\usage{
get_expressed_peaks_sce(peaks.sce.object, population.1,
  population.2 = NULL, threshold = 0.05)
}
\arguments{
\item{peaks.sce.object}{the peak-count SCE object}

\item{population.1}{target population}

\item{population.2}{background population If NULL (deafult) all non-population.1 cells}

\item{threshold}{percentage threshold of detected (non-zero) expression for including a peak}
}
\value{
an array of peak (or gene) names
}
\description{
Selects peaks that are considered expressed above some provided criteria within a target or
background cluster. Considers peaks expressed in some x\% of cells to be highly expressed. Returns the
union of peaks identified from the target and background cluster
}
\examples{
get_expressed_peaks_sce(peak.sce, "1")
get_expressed_peaks_sce(peak.sce, population.1 = "1", population.2 = "2")

}
