% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_usage.R
\name{get_expressed_peaks_seurat}
\alias{get_expressed_peaks_seurat}
\title{Identify highly expressed peaks}
\usage{
get_expressed_peaks_seurat(
  peaks.seurat.object,
  population.1,
  population.2 = NULL,
  threshold = 0.05
)
}
\arguments{
\item{peaks.seurat.object}{the peak-count Seurat object}

\item{population.1}{target cluster}

\item{population.2}{background cluster. If NULL (deafult) all non-target cells}

\item{threshold}{percentage threshold of detected (non-zero) expression for including a peak}
}
\value{
an array of peak (or gene) names
}
\description{
Selects peaks that are considered expressed above some provided criteria within a target or
background cluster. Considers peaks expressed in some x\% of cells to be highly expressed. Returns the
union of peaks identified from the target and background cluster
}
\examples{
\dontrun{
  get_highly_expressed_peaks(seurat.object, "1")
  get_highly_expressed_peaks(seurat.object, population.1 = "1", population.2 = "2")
}
}
