% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{get_highly_expressed_peaks_v2}
\alias{get_highly_expressed_peaks_v2}
\title{Identify highly expressed peaks}
\usage{
get_highly_expressed_peaks_v2(seurat.object, cluster1, cluster2 = NULL,
  threshold = 0.05)
}
\arguments{
\item{seurat.object}{the}

\item{cluster1}{target cluster}

\item{cluster2}{background cluster. If NULL (deafult) all non-target cells}

\item{threshold}{percentage threshold of detected (non-zero) expression for including a peak}
}
\value{
an array of peak (or gene) names
}
\description{
Selects peaks that are considered expressed above some provided criteria within a target or
background cluster. Considers peaks expressed in some x\% of cells to be highly expressed. Returns the
union of peaks identified from the target and background cluster
}
\examples{
get_highly_expressed_peaks(apa.seurat, "1")
get_highly_expressed_peaks(apa.seurat, cluster1 = "1", cluster2 = "2")

}
