% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_usage.R
\name{get_log2FC}
\alias{get_log2FC}
\title{Calculate Log2 fold-change}
\usage{
get_log2FC(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
log2 foldchange between x and y
}
\description{
Calculate Log2 fold-change between two vectors of log-normalised data (default Seurat normalisation)
}
