% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{get_log2FC_list}
\alias{get_log2FC_list}
\title{Calculate log2 fold-changes for a list of genes}
\usage{
get_log2FC_list(seurat.object, geneList, cluster1, cluster2 = NULL)
}
\arguments{
\item{seurat.object}{A polyA Seurat object}

\item{geneList}{list of genes to calculate log2 fold-changes for}

\item{cluster1}{target cluster}

\item{cluster2}{background cluster. If null uses all non-cluster1 cells.}
}
\value{
a list of log2 fold-changes
}
\description{
Get a list of Log2 fold-change values for a gene list between a specified cluster
and either all remaining cells (default) or an alternative cluster
}
\examples{
foldchange.list = get_log2FC_list(seurat.object, geneList, "1")

}
