% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{get_pvalue_list}
\alias{get_pvalue_list}
\title{Get a list of P-values for differential polyA testing}
\usage{
get_pvalue_list(expressionData, identities, cluster1, cluster2 = NULL,
  test.use = "MAST")
}
\arguments{
\item{expressionData}{a matrix of expression data}

\item{identities}{cell identity labels (normally clusters)}

\item{cluster1}{target cluster to test DE for}

\item{cluster2}{optional comparison cluster (all non-cluster1 cells by default)}

\item{test.use}{statistical test to use. Currently either MAST or t-test.}
}
\value{
a list of P-values
}
\description{
Given an expression matrix, a list of cluster identities and cluster labels, calculate
a list of differential expression P-values for the polyA sites in the expression matrix.
Currently can choose from two tests: MAST or t-test.
}
\examples{
p.values = get_pvalue_list(expressionData, identities, "1")

}
