% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_relative_expression_sce}
\alias{get_relative_expression_sce}
\title{Calculate relative expression between two or more peaks}
\usage{
get_relative_expression_sce(peaks.sce.object, peak.set = NULL,
  gene.name = NULL, feature.type = c("UTR3", "exon"))
}
\arguments{
\item{peaks.sce.object}{Seurat object}

\item{peak.set}{set of peaks}

\item{feature.type}{features to consider. 3'UTR and exon by default.}

\item{gene_name}{gene name for retrieving a set of peaks}
}
\value{
a matrix of relative expression
}
\description{
Calculate a relative expression between two or more peaks by dividing
the expression of each peak by the mean of the peak expression for that gene -
or set of provided peaks
}
\examples{
get_relative_expression(peaks.seurat, gene.name = "Cxcl12")

}
