% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_relative_expression_seurat}
\alias{get_relative_expression_seurat}
\title{Calculate relative expression between two or more peaks}
\usage{
get_relative_expression_seurat(
  peaks.seurat.object,
  peak.set = NULL,
  gene.name = NULL,
  feature.type = c("UTR3", "exon"),
  p.count = 1
)
}
\arguments{
\item{peaks.seurat.object}{Seurat object}

\item{peak.set}{set of peaks}

\item{gene.name}{gene name for retrieving a set of peaks}

\item{feature.type}{features to consider. 3'UTR and exon by default.}

\item{p.count}{Pseudo count}
}
\value{
a matrix of relative expression
}
\description{
Calculate a relative expression between two or more peaks by dividing
the expression of each peak by the mean of the peak expression for that gene -
or set of provided peaks
}
\examples{
\dontrun{
get_relative_expression_seurat(peaks.seurat.object, gene.name = "Cxcl12")
}
}
