% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_usage.R
\name{Log2ExpMean}
\alias{Log2ExpMean}
\title{Calculate average log2 expression}
\usage{
Log2ExpMean(x)
}
\arguments{
\item{x}{a vector of log-normalised data}
}
\value{
average value of the vector in log2-space
}
\description{
Calculate average expression of a peak/gene in log2 space
}
