% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{lr_test}
\alias{lr_test}
\title{logistic-regression test for differential expression}
\usage{
lr_test(f.labels, exp.matrix, value.labels, num_cells_fg, num_cells_bg,
  num_apa)
}
\arguments{
\item{f.labels}{class (cluster) labels for cells}

\item{exp.matrix}{matrix of counts}

\item{value.labels}{values}

\item{num_cells_fg}{number of 'positive' class cells}

\item{num_cells_bg}{number of background or 'negative' class cells}

\item{num_apa}{number of polyA sites used as predictors}
}
\value{
a P-value
}
\description{
Perform the logistic-regression test described by Lior Patcher and coleagues, which has
been adapted here for use on polyA sites. Builds a logistic regression
function using the polyA sites for a gene as the predictors. Performance
for correctly classifying class labels is compared to a null model to derive
a P-value for the gene.
}
\examples{
p.val = lr_test(f.labels, exp.matrix, value.labels, num_cells_fg, num_cells_bg, num_apa)

}
