% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{make_reference}
\alias{make_reference}
\title{Build gene start-end reference from a gtf file}
\usage{
make_reference(gtf_file)
}
\arguments{
\item{gtf_file}{gtf file

Takes a GTF file as input and creates a table of chromosome start-end
positions for each gene. Works with GTF files downloaded from 10x Genomics website.}
}
\description{
Build gene start-end reference from a gtf file
}
