% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{makeReference}
\alias{makeReference}
\title{Build gene star-end reference from a gtf file}
\usage{
makeReference(gtf_file)
}
\description{
Takes a GTF file as input and creates a table of chromosome start-end
positions for each gene. Works with GTF files downloaded from 10x Genomics website.
}
