% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{mast_de_test}
\alias{mast_de_test}
\title{Do MAST differential expression test.}
\usage{
mast_de_test(expressionData, identities, cluster1, cluster2 = NULL)
}
\arguments{
\item{expressionData}{a matrix of polyA site expression}

\item{identities}{cluster labels for cells}

\item{cluster1}{target cluser ID}

\item{cluster2}{comparison cluster ID}
}
\value{
a list of P-values
}
\description{
Apply a MAST differential expression test to polyA sites. Code derived from the Seurat 'MASTDETest'
function - see 'MASTDETest' at https://github.com/satijalab/seurat for more details.
}
\examples{
doMASTDETest(expressionData, identities, cluster1, cluster2)

}
