% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{NewPeakSCE}
\alias{NewPeakSCE}
\title{Create a new peak-counts single-cell experiment object from the peak counts}
\usage{
NewPeakSCE(peak.data, annot.info, cell.idents, tsne.coords = NULL,
  umap.coords = NULL, min.cells = 10, min.peaks = 200,
  norm.scale.factor = 10000, verbose = TRUE)
}
\arguments{
\item{peak.data}{matrix of peak counts}

\item{annot.info}{peak annotation information}

\item{cell.idents}{cell identities to be used for DU analysis}

\item{tsne.coords}{data-frame of t-SNE coordinates. Rownames should correspond to cell names.}

\item{umap.coords}{data-frame of UMAP coordinates. Rownames should correspond to cell names.}

\item{min.cells}{minimum number of cells for retaining a peak}

\item{min.peaks}{minimum number of peaks for retaining a cell}

\item{norm.scale.factor}{scale factor for log normalisation  function}
}
\value{
a new peak-level SCE object
}
\description{
Creates a new peak-counts single-cell experiment object from the peak counts and annotation table
}
\examples{
peak.sce = NewPeakSCE(apa.data, genes.seurat, annot.info)

}
