% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{NewPeakSeurat}
\alias{NewPeakSeurat}
\title{Create a new peak-level Seurat object from the peak counts}
\usage{
NewPeakSeurat(peak.data, annot.info, project.name = "PolyA",
  min.cells = 10, min.peaks = 200, norm.scale.factor = 10000)
}
\arguments{
\item{peak.data}{matrix of peak counts}

\item{annot.info}{peak annotation information}

\item{project.name}{project name passed to the Seurat object creation}

\item{min.cells}{minimum number of cells for retaining a peak}

\item{min.peaks}{minimum number of peaks for retaining a cell}

\item{norm.scale.factor}{scale factor for Seurat NormalizeData function}
}
\value{
a new peak-level Seurat object
}
\description{
Creates a new peak-level Seurat object from the peak counts and annotation table
}
\examples{
peak.seurat = NewPeakSeurat(peak.data, genes.seurat, annot.info)

}
