% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_tsne}
\alias{plot_tsne}
\title{t-SNE plot of cell populations}
\usage{
plot_tsne(seurat.object, col.set = NULL, title = NULL,
  do.plot = TRUE, pt.size = 0.75, show.labels = FALSE,
  return.plot = FALSE, simple.theme = FALSE)
}
\arguments{
\item{col.set}{a vector of colour codes corresponding to the number of clusters}

\item{title}{optional plot title}

\item{do.plot}{whether to print the plot to output (default: TRUE).}

\item{pt.size}{size of the point (default: 0.75)}

\item{show.labels}{whether to show the cluster labels (default: FALSE)}

\item{return.plot}{whether to return the ggplot object (default: FALSE)}

\item{simple.theme}{whether to remove axis (default: FALSE)}

\item{seurat.obect}{a Seurat object containing t-SNE coordinates and cluster ID's in @ident slot}
}
\value{
NULL by default. Returns a ggplot2 object if return.plot = TRUE
}
\description{
Based on t-SNE coordinates and cluster identities stored in a Seurat object,
plot t-SNE colouring cells according to cluster ID
}
\examples{
plot_tsne(apa.seurat, show.labels = TRUE)

}
