% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotRelativeExpressionTSNE}
\alias{PlotRelativeExpressionTSNE}
\title{Generate a t-SNE plot using relative expression}
\usage{
PlotRelativeExpressionTSNE(peaks.object, peaks.to.plot, do.plot = FALSE,
  figure.title = NULL, return.plot = TRUE, pt.size = 0.5,
  txt.size = 14)
}
\arguments{
\item{peaks.to.plot}{Set of peaks to plot}

\item{do.plot}{Whether to plot to output (TRUE by default)}

\item{figure.title}{Optional figure title}

\item{pt.size}{size of the points on the t-SNE plot}

\item{seurat.object}{Seurat object}
}
\value{
a ggplot2 object
}
\description{
Given two or more peaks to plot, a relative expression score and
plot on t-SNE coordinates
}
\examples{
PlotRelativeExpressionTSNE(peaks.seurat, this.peak.set)

}
