% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotRelativeExpressionUMAP}
\alias{PlotRelativeExpressionUMAP}
\title{Generate a t-SNE plot using relative expression}
\usage{
PlotRelativeExpressionUMAP(
  peaks.object,
  peaks.to.plot,
  do.plot = FALSE,
  figure.title = NULL,
  return.plot = TRUE,
  pt.size = 0.5,
  txt.size = 14
)
}
\arguments{
\item{peaks.object}{Seurat object}

\item{peaks.to.plot}{Set of peaks to plot}

\item{do.plot}{Whether to plot to output (TRUE by default)}

\item{figure.title}{Optional figure title}

\item{return.plot}{Boolean of whether to return plot (default TRUE)}

\item{pt.size}{size of the points on the t-SNE plot. Default 0.5}

\item{txt.size}{size of text. Default 14}
}
\value{
a ggplot2 object
}
\description{
Given two or more peaks to plot, calculate a relative expression score and
plot on UMAP coordinates
}
\examples{
\dontrun{
   PlotRelativeExpressionUMAP(peaks.seurat, this.peak.set)
 }
}
