% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{polya_seurat_from_gene_object}
\alias{polya_seurat_from_gene_object}
\title{Create a polyA Seurat object using a gene-level object}
\usage{
polya_seurat_from_gene_object(peak.data, genes.seurat, annot.info,
  project.name = "PolyA", min.cells = 10, min.peaks = 200,
  norm.scale.factor = 10000)
}
\arguments{
\item{peak.data}{matrix of peak counts}

\item{genes.seurat}{a Seurat object}

\item{annot.info}{peak annotation information}

\item{project.name}{project name passed to the Seurat object creation}

\item{min.cells}{minimum number of cells for retaining a peak}

\item{min.peaks}{minimum number of peaks for retaining a cell}

\item{norm.scale.factor}{scale factor for Seurat NormalizeData function}
}
\value{
a new polyA-level Seurat object
}
\description{
Creates a new polyA Seurat object, importing information on clustering and dimensionality reduction,
such as t-SNE coordinates, from a Seurat object that has been processed at the gene level.
}
\examples{
apa.seurat = polya_seurat_from_gene_object(apa.data, genes.seurat, annot.info)

}
