% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{readMEX}
\alias{readMEX}
\title{Read in peak data saved in MEX format}
\usage{
readMEX(mm.file, barcodes.file, genes.file)
}
\arguments{
\item{mm.file}{count matrix in MEX format}

\item{barcodes.file}{file containing cell barcodes corresponding to columns in the matrix}

\item{genes.file}{file containing peak or gene names corresponding to rows in the matrix}
}
\value{
a sparseMatrix
}
\description{
Read in peak data saved in MEX format
}
\examples{
count.mat = readMEX(mm.file, barcodes.file, genes.file)

}
