% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{relative_expression_tsne}
\alias{relative_expression_tsne}
\title{Generate a t-SNE plot using relative expression}
\usage{
relative_expression_tsne(seurat.object, peaks.to.plot, do.plot = TRUE,
  figure.title = NULL, return.plot = TRUE, pt.size = 0.5)
}
\arguments{
\item{seurat.object}{Seurat object}

\item{peaks.to.plot}{Set of peaks to plot}

\item{do.plot}{Whether to plot to output (TRUE by default)}

\item{figure.title}{Optional figure title}

\item{pt.size}{size of the points on the t-SNE plot}
}
\value{
a ggplot2 object
}
\description{
Given two or more peaks to plot,
}
\examples{
relative_expression_tsne(peaks.seurat, this.peak.set)

}
