% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{select_gene_polyas}
\alias{select_gene_polyas}
\title{return polyAs associated with a select gene}
\usage{
select_gene_polyas(apa.seurat.object, gene, feature.type = c("UTR3",
  "UTR5", "exon", "intron"))
}
\arguments{
\item{apa.seurat.object}{Seurat polyA object}

\item{gene}{Gene name}

\item{feature.type}{type of genomic features to use}
}
\value{
a list of polyA IDs
}
\description{
return polyAs associated with a select gene
}
\examples{
polya.list = select_gene_polyas(apa.seurat, "PTPRC", feature.type = c("UTR3", "exon"))

}
