% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{SelectGenePeaks}
\alias{SelectGenePeaks}
\title{Return peaks associated with a select gene.}
\usage{
SelectGenePeaks(peaks.object, gene, feature.type = c("UTR3", "UTR5",
  "exon", "intron"))
}
\arguments{
\item{peaks.object}{Peaks SCE or Seurat object.}

\item{gene}{Gene name}

\item{feature.type}{type of genomic features to use}
}
\value{
a list of peak IDs
}
\description{
Returns peaks associated with a select gene.
}
\examples{
peak.list = SelectGenePeaks(peaks.object, "PTPRC", feature.type = c("UTR3", "exon"))

}
