% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBams.R
\name{seqmonk_to_rle}
\alias{seqmonk_to_rle}
\title{Example of how to use function:}
\usage{
seqmonk_to_rle(df, col_idx = 13:28)
}
\description{
library("data.table")
endothelial_cov <- fread(file="c:/BAM/Harvey/scpolyA/Porrello_Support_Files/Porrello_Endothelial.F-CycCl_vs_F-Act.wig.txt.gz", sep = "\t", header = TRUE)
EC_coverage.rle <- seqmonk_to_rle(endothelial_cov, col_idx = 13:28)
}
