% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.R
\name{BaseComposition}
\alias{BaseComposition}
\title{Identify polyA motif and/or polyA stretches from provided genomic coordinates}
\usage{
BaseComposition(genome = NULL, chrom = NULL, start = NULL,
  stop = NULL, strand = NULL, coord = NULL, offset = -50,
  length = 250, mismatch = 1, AT_length = 13)
}
\arguments{
\item{genome}{genome object of organism.}

\item{chrom}{chromosome}

\item{start}{Upstream start position}

\item{strand}{'+' or '-'. This will define the applied direction of offset}

\item{coord}{coordinates}

\item{offset}{The}

\item{length}{How many nucleotides of DNA sequence to return}

\item{mismatch}{: The max number of mismatches allowed in poly A/T stretch (default 1)}

\item{AT_length}{: length of A/T to search for within input sequence (default 13)}

\item{end}{downstream end position}
}
\value{
a dataframe with appended columns containing annotation

chrom <- 'chr16'
start <- 49896378
stop  <- 49911102
strand <- '+'
genome <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

# Tmem126a intronic peak that coincides with a poly(A) rich region.
chrom  <- 'chr7'
start  <- 90451180
stop   <- 90451380
strand <- '-'
output <-BaseComposition(genome=genome, chrom=chrom, start=start, stop=stop, strand=strand)

# Dync1h1 intronic peak that coincides with a long poly(T) rich region
coord <- "chr12:110609400-110609800:1"
output <-BaseComposition(genome=genome, coord=coord)

TO DO:
* If peak falls at end of exon then need to obtain sequence from next exon. This
would require passing exon junction information.
}
\description{
chrom <- chr8
}
\details{
start <- 70331172

stop <- 70331574

strand <- "+"

You need to run the following code:
     genome <-  GenomicFeatures::makeTxDbFromGFF(gtf_file, format="gtf")

annotationType can be c("any", "start", "end", "within", "equal"),
}
