% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{CountPeaks}
\alias{CountPeaks}
\title{count polyA sites in cells}
\usage{
CountPeaks(peak.sites.file, gtf.file, bamfile, whitelist.file, output.file,
  countUMI = TRUE, ncores = 1)
}
\arguments{
\item{peak.sites.file}{a file containing peak coordinates generated by FindPeaks}

\item{gtf.file}{reference (GTF) file}

\item{bamfile}{scRNA-seq BAM file}

\item{whitelist.file}{white list file}

\item{output.file}{name of output}

\item{countUMI}{whether to count UMIs (default: TRUE)}
}
\value{
NULL. Writes counts to file.
}
\description{
Generates a count matrix
}
\examples{
CountPeaks(peak.sites.file, reference.file, bamfile, whitelist.file, output.file)

}
