% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_polyA.R
\name{FindPeaks}
\alias{FindPeaks}
\title{Do peak calling on a scRNA-seq BAM file}
\usage{
FindPeaks(
  output.file,
  gtf.file,
  bamfile,
  junctions.file,
  min.jcutoff = 50,
  min.jcutoff.prop = 0.05,
  min.cov.cutoff = 500,
  min.cov.prop = 0.05,
  min.peak.cutoff = 200,
  min.peak.prop = 0.05,
  ncores = 1,
  chr.names = NULL,
  filter.chr = TRUE
)
}
\arguments{
\item{output.file}{a file containing polyA sites}

\item{gtf.file}{reference (GTF) file}

\item{bamfile}{scRNA-seq BAM file}

\item{junctions.file}{file of splice junctions (e.g. produced by regtools)}

\item{min.jcutoff}{minimum number of spliced reads across a junction for it to be considered (default: 50).}

\item{min.jcutoff.prop}{minimum proportion of junction reads out of all junction reads for that gene (default: 0.05)}

\item{min.cov.cutoff}{min.cov.cutoff}

\item{min.cov.prop}{min.cov.prop}

\item{min.peak.cutoff}{min.peak.cutoff}

\item{min.peak.prop}{min.peak.prop}

\item{ncores}{number of cores to use}

\item{chr.names}{names of chromosomes}

\item{filter.chr}{names of chromosomes to filter}
}
\value{
NULL. Writes counts to file.
}
\description{
Do peak calling on a scRNA-seq BAM file...
}
\examples{

extdata_path <- system.file("extdata",package = "Sierra")
reference.file <- paste0(extdata_path,"/Vignette_cellranger_genes_subset.gtf")
junctions.file <- paste0(extdata_path,"/Vignette_example_TIP_sham_junctions.bed")

bamfile <- c(paste0(extdata_path,"/Vignette_example_TIP_sham.bam"),
             paste0(extdata_path,"/Vignette_example_TIP_mi.bam") )
             
peak.output.file <- c("Vignette_example_TIP_sham_peaks.txt",  "Vignette_example_TIP_MI_peaks.txt")


FindPeaks(output.file=peak.output.file[1], gtf.file = reference.file, 
             bamfile=bamfile[1], junctions.file=junctions.file)

}
