% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bams.R
\name{geneToGR}
\alias{geneToGR}
\title{geneToGR converts a gene sybol to genomic ranges coordinate}
\usage{
geneToGR(geneID, gtf_gr)
}
\arguments{
\item{geneID}{: Gene symbol}

\item{gtf_gr}{: Granges object of a gtf file

gtf_file <- "u:/Reference/mm10/cellranger_genes.gtf.gz"
gtf_gr <- rtracklayer::import(gtf_file)
gtf_gr=gtf_gr, geneSymbol="Dnajc19"
geneGR  <- geneToGR(gtf_gr=gtf_gr, geneSymbol="Dnajc19")}
}
\description{
geneToGR converts a gene sybol to genomic ranges coordinate
}
