% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBams.R
\name{geneToGR}
\alias{geneToGR}
\title{geneToGR converts a gene sybol to genomic ranges coordinate}
\usage{
geneToGR(geneID, gtf_gr)
}
\arguments{
\item{geneID}{: Gene symbol}

\item{gtf_gr}{: Gragnes object of a gtf file}
}
\description{
geneToGR converts a gene sybol to genomic ranges coordinate
}
