% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBams.R
\name{merge_bam_coverage}
\alias{merge_bam_coverage}
\title{merge_bam_coverage}
\usage{
merge_bam_coverage(bamfiles, to_extract)
}
\arguments{
\item{bamfile}{: A list of BAM files that are to be merged}
}
\description{
merge_bam_coverage
}
