% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_merging.R
\name{MergePeakCoordinates}
\alias{MergePeakCoordinates}
\title{Merge peaks across a list of data-sets}
\usage{
MergePeakCoordinates(peak.dataset.table, output.file, sim.thresh = 0.75,
  allow.match.var = 0.25, ncores = 1)
}
\arguments{
\item{peak.dataset.table}{a dataframe with two required columnss: one called "Peak_file" , which
contains file names of the peak data-sets to be merged and labels ("Identifier") for each file.}

\item{output.file}{file to write the set of merged peaks to}

\item{sim.thresh}{The required similarity threshold for merging (default: 0.75)}

\item{allow.match.var}{The allowance for deviation from the sim.thresh for comparison peaks (default: 0.25)}

\item{ncores}{number of cores to use (default 1)}
}
\value{
NULL. writes out a set of merged peaks to output.file
}
\description{
Takes as input a list of named peaks obtained from running peak calling on multiple data-sets.
First goes through each peak set and check what peaks within each set should be merged (self-merging).
Merging is based on similarity criteria set by sim.thresh and allow.match.var.
Then compares each peak set as a reference to the remaining sets to identify peaks that should be merged.
Returns a list of peaks that have been merged, as well as the unique peaks from each data-set.
}
\examples{
MergePeakCoordinates(peak.dataset.table, output.file, ncores = 4)

}
