% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{NewPeakSeurat}
\alias{NewPeakSeurat}
\title{Create a new peak-level Seurat object from the peak counts}
\usage{
NewPeakSeurat(
  peak.data,
  annot.info,
  project.name = "PolyA",
  cell.idents = NULL,
  tsne.coords = NULL,
  umap.coords = NULL,
  min.cells = 10,
  min.peaks = 200,
  norm.scale.factor = 10000,
  filter.gene.mismatch = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{peak.data}{matrix of peak counts}

\item{annot.info}{peak annotation information}

\item{project.name}{project name passed to the Seurat object creation}

\item{cell.idents}{a list of cell identities (optional)}

\item{tsne.coords}{a data-frame of t-SNE coordinates (optional)}

\item{umap.coords}{a data-frame of UMAP coordinates (optional)}

\item{min.cells}{minimum number of cells for retaining a peak}

\item{min.peaks}{minimum number of peaks for retaining a cell}

\item{norm.scale.factor}{scale factor for Seurat NormalizeData function}

\item{filter.gene.mismatch}{whether to filter out peaks with ambiguous gene mappings}

\item{verbose}{whether to print output}
}
\value{
a new peak-level Seurat object
}
\description{
Creates a new peak-level Seurat object from the peak counts and annotation table
}
\examples{

\dontrun{ 
     peak.seurat = NewPeakSeurat(peak.data, genes.seurat, annot.info)
 }
}
