% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotCoverage}
\alias{PlotCoverage}
\title{plotCoverage}
\usage{
PlotCoverage(genome_gr, geneSymbol = "", wig_data = NULL,
  bamfiles = NULL, wig_same_strand = TRUE, genome = NULL,
  pdf_output = FALSE, wig_data.tracknames = NULL,
  bamfile.tracknames = NULL, output_file_name = "",
  zoom_3UTR = FALSE)
}
\arguments{
\item{wig_data}{can be a data frame or a genomic ranges object. Must be stranded.}

\item{bamfiles}{: BAM filenames that are to be displayed as data tracks}

\item{wig_same_strand}{Display same strand or opposing strand of wig data (compared to reference gene)}

\item{pdf_output}{: If true will create output pdf files}

\item{wig_data.tracknames}{: WIG track display names. Assumed to be in same order as wig_data.}

\item{bamfile.tracknames}{: BAM track display names. Assumed to be in same order as bamfiles.}

\item{output_file_name}{: Used if pdf_output is true. Location of where files will be placed.}

\item{zoom_UTR}{: If TRUE will create a second figure which will zoom in on 3'UTR.}
}
\value{
NULL by default.
}
\description{
Plots read coverage across a gene for a set of BAM files and/or wig data.
}
\examples{

extdata_path <- system.file("extdata",package = "Sierra")
reference.file <- paste0(extdata_path,"/Vignette_cellranger_genes_subset.gtf")
gtf_gr <- rtracklayer::import(reference.file)
bam.files <- c(paste0(extdata_path,"/Fibroblast.Cxcl12.bam"),
                 paste0(extdata_path,"/EC.Cxcl12.bam"))

PlotCoverage(genome_gr = gtf_gr, geneSymbol = "Cxcl12", genome = "mm10", 
           bamfiles = bam.files, bamfile.tracknames=c("Fibroblast", "Endothelial"))

}
