% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotCoverage}
\alias{plotCoverage}
\title{plotCoverage}
\usage{
plotCoverage(genome_gr, geneSymbol = "", wig_data = NULL,
  bamfiles = NULL, wig_same_strand = TRUE, genome = NULL,
  pdf_output = FALSE, output_file_name = "", zoom_3UTR = FALSE)
}
\arguments{
\item{wig_data}{can be a data frame or a genomic ranges object. Must be stranded.}

\item{bamfiles}{: BAM filenames that are to be displayed as data tracks}

\item{wig_same_strand}{Display same strand or opposing strand of wig data (compared to reference gene)}

\item{pdf_output}{: If true will create output pdf files}

\item{output_file_name}{: Used if pdf_output is true. Location of where files will be placed.}

\item{zoom_UTR}{: If TRUE will create a second figure which will zoom in on 3'UTR.}
}
\value{
NULL by default.
}
\description{
plotCoverage
}
\examples{

gtf_file <- "u:/Reference/mm10/cellranger_genes.gtf.gz"
gtf_gr <- rtracklayer::import(gtf_file)

endothelial_cov <- read.table(file="c:/BAM/Harvey/scpolyA/Porrello_Support_Files/Porrello_Endothelial.F-CycCl_vs_F-Act.wig.txt.gz", sep = "\\t", header = TRUE)
df <- endothelial_cov[,c("Chromosome","Start","End","Probe.Strand", "MIP1_1.BAM")]
df <- cbind(endothelial_cov[,c("Chromosome","Start","End","Probe.Strand")],endothelial_cov[, 13:15])
colnames(df)[1:4] <- c("chrom", "start","end", "strand")
wig_data<- GenomicRanges::makeGRangesFromDataFrame(df,keep.extra.columns=TRUE)
 plotCoverage(genome_gr=gtf_gr, geneSymbol="Prkar1a", wig_data= wig_data)

plotCoverage(genome_gr=gtf_gr, geneSymbol="Dnajc19", bamfiles = "c:/TEMP/Bams/F-SH.Dnajc19.bam",wig_data= wig_data)

}
