% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotRelativeExpressionBox}
\alias{PlotRelativeExpressionBox}
\title{Generate a box plot plot using relative expression}
\usage{
PlotRelativeExpressionBox(
  peaks.object,
  peaks.to.plot,
  do.plot = FALSE,
  figure.title = NULL,
  return.plot = TRUE,
  pt.size = 0.5,
  col.set = NULL,
  txt.size = 14
)
}
\arguments{
\item{peaks.object}{Peak object of either Seurat or SCE class}

\item{peaks.to.plot}{Set of peaks to plot}

\item{do.plot}{Whether to plot to output (TRUE by default)}

\item{figure.title}{Optional figure title}

\item{return.plot}{Boolean (default True) identifying if plot should be returned.}

\item{pt.size}{Size of the points on the t-SNE plot (default 0.5)}

\item{col.set}{col set (default NULL)}

\item{txt.size}{sie of text (default 14)}
}
\value{
a ggplot2 object
}
\description{
Given two or more peaks to plot, a relative expression score and
generate a box plot according to cell identities
}
\examples{
\dontrun{
   PlotRelativeExpressionBox(peaks.object, this.peak.set)
 }
}
