% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotRelativeExpressionTSNE}
\alias{PlotRelativeExpressionTSNE}
\title{Generate a t-SNE plot using relative expression}
\usage{
PlotRelativeExpressionTSNE(
  peaks.object,
  peaks.to.plot,
  do.plot = FALSE,
  figure.title = NULL,
  return.plot = TRUE,
  pt.size = 0.5,
  txt.size = 14
)
}
\arguments{
\item{peaks.object}{peak object either Seurat or SingleCellExperiment class}

\item{peaks.to.plot}{Set of peaks to plot}

\item{do.plot}{Whether to plot to output (TRUE by default)}

\item{figure.title}{Optional figure title}

\item{return.plot}{boolean of whether to return plot. Default is TRUE.}

\item{pt.size}{size of the points on the t-SNE plot. Default 0.5}

\item{txt.size}{size of text. Default 14}
}
\value{
a ggplot2 object
}
\description{
Given two or more peaks to plot, a relative expression score and
plot on t-SNE coordinates
}
\examples{

## Load example data for two peaks from the Cxcl12 gene
extdata_path <- system.file("extdata",package = "Sierra")
load(paste0(extdata_path, "/Cxcl12_example.RData"))
load(paste0(extdata_path, "/TIP_cell_info.RData"))

## Create an SCE object holding the peak data
peaks.sce <- NewPeakSCE(peak.data = peak.counts, 
                        annot.info = peak.annotations, 
                        cell.idents = tip.populations, 
                        tsne.coords = tip.tsne.coordinates,
                        min.cells = 0, min.peaks = 0)
                        
## Plot relative expression of example peaks on t-SNE coordinates
PlotRelativeExpressionTSNE(peaks.object = peaks.sce, 
      peaks.to.plot = c("Cxcl12:6:117174603-117175050:1", "Cxcl12:6:117180974-117181367:1"))
 
}
