% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PlotRelativeExpressionViolin}
\alias{PlotRelativeExpressionViolin}
\title{Generate a violin plot plot using relative expression}
\usage{
PlotRelativeExpressionViolin(peaks.object, peaks.to.plot, do.plot = TRUE,
  figure.title = NULL, return.plot = TRUE, pt.size = 0.5,
  col.set = NULL, txt.size = 14, add.jitter = TRUE,
  jitter.pt.size = 0.25)
}
\arguments{
\item{peaks.object}{Peak object of either Seurat or SCE class}

\item{peaks.to.plot}{Set of peaks to plot}

\item{do.plot}{Whether to plot to output (TRUE by default)}

\item{figure.title}{Optional figure title}

\item{pt.size}{size of the points on the t-SNE plot}

\item{add.jitter}{whether to add a geom_jitter to the plot (default: TRUE)}

\item{jitter.pt.size}{size of point for geom_jitter (default = 0.25)}
}
\value{
a ggplot2 object
}
\description{
Given two or more peaks to plot, a relative expression score and
generate a violin plot according to cell identities
}
\examples{
PlotRelativeExpressionViolin(peaks.object, this.peak.set)

}
