% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBams.R
\name{rle_to_WIG}
\alias{rle_to_WIG}
\title{load(file="c:/BAM/scRNA_polyA/FC.RData")
gtf_file <- "u:/Reference/mm10/cellranger_genes.gtf.gz"
gtf_gr <- rtracklayer::import(gtf_file)}
\usage{
rle_to_WIG(rle_input, gtf_gr = gtf_gr, geneSymbol = "Dnajc19")
}
\description{
load(file="c:/BAM/scRNA_polyA/FC.RData")
gtf_file <- "u:/Reference/mm10/cellranger_genes.gtf.gz"
gtf_gr <- rtracklayer::import(gtf_file)
}
