% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_util.R
\name{SelectGenePeaks}
\alias{SelectGenePeaks}
\title{Return peaks associated with a select gene.}
\usage{
SelectGenePeaks(
  peaks.object,
  gene,
  feature.type = c("UTR3", "UTR5", "exon", "intron")
)
}
\arguments{
\item{peaks.object}{Peaks SCE or Seurat object.}

\item{gene}{Gene name}

\item{feature.type}{type of genomic features to use}
}
\value{
a list of peak IDs
}
\description{
Returns peaks associated with a select gene.
}
\examples{


#' extdata_path <- system.file("extdata",package = "Sierra")
load(paste0(extdata_path, "/Cxcl12_example.RData"))
load(paste0(extdata_path, "/TIP_cell_info.RData"))

## Create an suerat object holding the peak data
peaks.seurat <- NewPeakSeurat(peak.data = peak.counts, 
                        annot.info = peak.annotations, 
                        cell.idents = tip.populations, 
                        tsne.coords = tip.tsne.coordinates,
                        min.cells = 0, min.peaks = 0)

peak.list <- SelectGenePeaks(peaks.object =  peaks.seurat ,gene = "Cxcl12")

}
