% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBams.R
\name{splitBam}
\alias{splitBam}
\title{splitBam}
\usage{
splitBam(bam, cellbc.df, outdir, yieldSize = 1e+06)
}
\arguments{
\item{bam}{CellRanger outputted bam file with the CB field}

\item{cellbc.df}{data frame of the cell barcode, needs to have the column names: "celltype" and "cellbc"}

\item{outdir}{directory to output the bam files. The bam files will be called [celltype].bam}

\item{yieldSize}{number of lines of bam files to load. Default: 1000000}
}
\value{
load(paste(extdata_path,"TIP_vignette_gene_Seurat.RData",sep="/"))
cellbc.df <- data.frame(celltype=genes.seurat@active.ident, cellbc= names(genes.seurat@active.ident))
bam <- "c:/BAM/Harvey/scpolyA/one_percent.bam"
splitBam(bam, cellbc.df, "c:/TEMP/")
}
\description{
Utility to split a bam file into multiple bam files based on the barcode
}
\details{
Given a bam file that was processed by CellRanger, splitBam splits the 
bam into multiple bam files, one per cell barcode. 
Bam file needs to have the barcode stored in the "CB" field.
}
