% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bams.R
\name{SplitBam}
\alias{SplitBam}
\title{Utility to split a bam file into multiple bam files based on the barcode}
\usage{
SplitBam(
  bam,
  cellbc.df,
  outdir = NULL,
  yieldSize = 1e+06,
  gtf_gr = NULL,
  geneSymbol = NULL,
  gi_ext = 50,
  rle_output = FALSE,
  exportFastqHeader = FALSE
)
}
\arguments{
\item{bam}{CellRanger outputted bam file with the CB field}

\item{cellbc.df}{data frame of the cell barcode, needs to have the column names: "celltype" and "cellbc"}

\item{outdir}{directory to output the bam files. The bam files will be called [celltype].bam. If NULL no BAM file created.}

\item{yieldSize}{number of lines of bam files to load. Default: 1000000}

\item{gtf_gr}{gene model genomic ranges. Only used if geneSymbol is defined.}

\item{geneSymbol}{Gene symbol. Used to identify the genomic coordinates to extract reads from.}

\item{gi_ext}{The number of nucleotides to extend the genomic interval in extracting reads from (default 50).}

\item{rle_output}{If TRUE will generate and return rle_list object}

\item{exportFastqHeader}{If TRUE will generate a txt output file that has same prefix as bam file containing fastq header IDs}
}
\value{
a rleList of coverage for each cell type
}
\description{
Given a bam file that was processed by CellRanger, splitBam splits the
bam into multiple bam files, one per cell barcode.
Bam file needs to have the barcode stored in the "CB" field.
}
\examples{
library('Sierra')
\dontrun{
extdata_path <- system.file("extdata",package = "scpolya")
load(paste(extdata_path,"TIP_vignette_gene_Seurat.RData",sep="/"))
cellbc.df <- data.frame(celltype=genes.seurat@active.ident, 
                        cellbc= names(genes.seurat@active.ident))
bamfile <- c(paste0(extdata_path,"/Vignette_example_TIP_sham.bam")

SplitBam(bam, cellbc.df)
}

# Example 2 extract reads that overlap a gene

extdata_path <- system.file("extdata",package = "Sierra")
gtf.file <- paste0(extdata_path,"/Vignette_cellranger_genes_subset.gtf")
gtf.gr <- rtracklayer::import(gtf.file)

load(paste(extdata_path,"TIP_vignette_gene_Seurat.RData",sep="/"))
cellbc.df <- data.frame(celltype=genes.seurat@active.ident, 
                       cellbc= names(genes.seurat@active.ident))
                       
                       
bam.file <- paste0(extdata_path,"/Vignette_example_TIP_mi.bam")
outdir <-  tempdir()  # change this to a meaningful location
SplitBam(bam.file, cellbc.df, outdir=outdir, gtf_gr=gtf.gr, geneSymbol="Dnajc19")


}
